refactor: separate rust API from wasm API
