+++ .github/workflows/ci.yml
@@ -53,12 +53,12 @@ jobs:
           cargo test --locked --release --all-features --all-targets
           deno task test
 
-      # - name: 📦  publish
-      #   if: |
-      #     contains(matrix.os, 'ubuntu') &&
-      #     github.repository == 'denoland/deno_emit' &&
-      #     startsWith(github.ref, 'refs/tags/')
-      #   env:
-      #     CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
-      #   run: |
-      #     cargo publish -vv
+      - name: 📦  publish
+        if: |
+          contains(matrix.os, 'ubuntu') &&
+          github.repository == 'denoland/deno_emit' &&
+          startsWith(github.ref, 'refs/tags/')
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+        run: |
+          cargo publish -p deno_emit -vv
