+++ .gitignore
@@ -1,2 +1,4 @@
 /target
+/rs-lib/target
+/wasm/target
 .vscode/
