+++ rs-lib/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "deno_emit"
+version = "0.0.2"
+edition = "2021"
+description = "module transpiling and emitting for deno"
+homepage = "https://deno.land/"
+repository = "https://github.com/denoland/deno_emit"
+authors = ["the Deno authors"]
+license = "MIT"
+
+[dependencies]
+anyhow = "1.0.44"
+base64 = "0.13.0"
+deno_ast = { version = "0.14.0", features = ["bundler", "codegen", "module_specifier", "proposal", "react", "sourcemap", "transforms", "typescript", "visit", "transpiling"] }
+deno_graph = "0.26.0"
+futures = "0.3.17"
+js-sys = { version = "0.3.55", optional = true }
+parking_lot = { version = "0.11.2", features = ["wasm-bindgen"] }
+serde = { version = "1.0.130", features = ["derive", "rc"] }
+serde_json = { version = "1.0.68", features = ["preserve_order"] }
