+++ rs-lib/src/bundle_hook.rs
