+++ rs-lib/src/emit.rs
