+++ rs-lib/src/text.rs
